Security Implementation stage 2

+++ src/main/java/com/hudyma/CarRental2024/repository/TokenRepository.java
@@ -12,7 +12,7 @@ public interface TokenRepository extends JpaRepository <Token, Integer> {
     @Query(value = "select t from Token t inner join User u on " +
             "t.user.id = u.id where u.id = :id and " +
             "(t.expired = false or t.revoked = false)")
-    List<Token> findAllValidTokensByUserId (Integer id);
+    List<Token> findAllValidTokensByUserId (Long id);
 
     Optional<Token> findByToken (String token);
 }
